Added entry in manifest for gradle_command.
Cargo apk wouldn't build my project on windows, because the gradle in my
path was a .bat file. On the command line, I could execute it as
'gradle', but when using rust's 'Command' object it had to be .gradle,
otherwise a 'file not found' error was returned.

I added an optional gradle_command field to allow me to specify
gradle.bat in my projects.